+++ examples/vite-sub/vite.config.ts
@@ -8,8 +8,8 @@ import Components from 'unplugin-vue-components/vite'
 import { ElementPlusResolver } from 'unplugin-vue-components/resolvers'
 import { visualizer } from 'rollup-plugin-visualizer'
 import Inspect from 'vite-plugin-inspect'
+// import Legacy from '@vitejs/plugin-legacy'
 import config from './dubhe.config'
-
 export const pubConfig = {
   project: 'vitesub',
   entry: {
@@ -39,6 +39,7 @@ export default (): UserConfig => {
     },
     plugins: [
       Inspect(),
+      // Legacy(),
       visualizer(),
       vue(),
       AutoImport({